Apply app bar theme to request detail dialog
