+++ src/utils/AppBar.tsx
@@ -1,0 +1,66 @@
+//
+// AppBar.tsx
+//
+// Copyright (c) 2022 Hironori Ichimiya <hiron@hironytic.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+
+import { AppBar as MUIAppBar, AppBarProps, createTheme, ThemeProvider } from "@mui/material";
+import React from "react";
+
+const appBarTheme = createTheme({
+  palette: {
+    primary: {
+      main: "#f5c300",
+      contrastText: "#fff",
+    },
+  },
+  components: {
+    MuiAppBar: {
+      styleOverrides: {
+        root: {
+          backgroundColor: "#11509a",
+        },
+      },
+    },
+    MuiTab: {
+      styleOverrides: {
+        root: {
+          color: "white",
+          opacity: 0.7,
+          "&.Mui-selected": {
+            color: "white",
+            opacity: 1.0,
+          },
+        },
+      },
+    },
+  },
+});
+
+export function AppBar({ position, children }: AppBarProps): JSX.Element {
+  return (
+    <ThemeProvider theme={appBarTheme}>
+      <MUIAppBar position={position}>
+        {children}
+      </MUIAppBar>
+    </ThemeProvider>
+  )
+}
